Bugfix to posthooks and partial decoupling (#25)
* requirements clean up

* toml setup; docker build for observer and dummy nodes

* docker observer

* move observer and dummy docker setup into one file

* docker does not need reqs as tracr installs them

* tracr doesn't get installed with dockerfile; dummy node dockerfile needs torch install

* docker installs tracr from local

* deploy node with cuda/arch args

* dockerignroe to not include outside src/ files

* basic tests for tracr

* run built-in alexnet test from deploy_node.sh

* docker runs builtin alexnet test

* entry point for torch install

* dynamic torch install for test node

* dos2unix shell scripts when running observer at entrypoint

* toml file updates

* dos2unix at docker build

* set up run.sh to handle experiment run command

* #TODO: fix run.sh for experiment run

* #TODO: update entrypoint and run

* alexnetsplit yaml file updated

* entrypoint now actually starts the nodes

* Configure .sh files to have LF line endings

* Apply line ending normalization

* Separate dockerfiles for observer and participant

* experiment design base.py dependency on torch needs to be fixed

* fix docker referencing files

* separate reqs for observer and participant

* drop torch from typing

* cleanup

* bugfixes

---------
