+++ .bashrc
@@ -1,0 +1,1 @@
+alias tracr_run="bash run.sh"
