+++ .gitattributes
@@ -1,1 +1,2 @@
-* text eol=lf
+# Ensure all shell scripts have LF (Unix) line endings
+*.sh text eol=lf
