--- AppData/known_devices.yaml.example
@@ -1,42 +1,0 @@
-# This file is used to store information about the devices that have been
-# configured for use with experiments. In the near future, there will be no
-# need to edit this file manually.
-#
-# Each device has:
-#   1. a unique nickname (the top-level keys in this file)
-#   2. one of the supported device types
-#   3. a list of connection parameters, each of which has:
-#       a. a hostname or IP address (host)
-#       b. a username (user)
-#       c. a path to a private key file, which should be stored in the
-#          "pkeys" directory under "AppData" (pkey_fp) *
-#       d. a boolean indicating whether this is the default connection
-#          parameter for this device (default) **
-#
-#  * The path is used within the docker container, so it shouldn't be changed
-#    to an acutal path on the host machine. See the volume mapping in run.sh
-#    if this doesn't make sense.
-# ** There may be multiple connection parameters for a device if, for instance,
-#    the device is accessible from the local network and from the internet; or
-#    if there are different users for the same device.
-
-
-HomePi4:
-  device_type: raspberrypi-4b
-  connection_params:
-    - host: pi-4.fios-router.home
-      user: joe22
-      pkey_fp: /usr/src/tracr/AppData/pkeys/id_rsa
-      default: true
-    - host: home-pi.duckdns.org
-      user: joe22
-      pkey_fp: /usr/src/tracr/AppData/pkeys/id_rsa
-      default: false
-
-RACRJnano4gb-1:
-  device_type: jetson-nano-4gb
-  connection_params:
-    - host: 192.168.1.212
-      user: racr
-      pkey_fp: /usr/src/tracr/AppData/pkeys/id_rsa
-      default: true
