+++ Dockerfile.observer
@@ -1,0 +1,28 @@
+# Use an official Python runtime as a parent image
+FROM python:3.11.7
+
+# Set the working directory
+WORKDIR /app/
+
+# Install system dependencies
+RUN apt-get update && \
+    apt-get install -y libgl1-mesa-glx && \
+    rm -rf /var/lib/apt/lists/*
+
+# Copy the requirements file
+COPY requirements.observer.txt .
+
+# Install Python dependencies
+RUN pip install --upgrade pip setuptools wheel
+RUN pip install -r requirements.observer.txt
+
+# Copy the entire project into the container
+COPY . /app/
+
+# Set PYTHONPATH
+# ENV PYTHONPATH /app/src/
+
+# Expose the necessary ports
+EXPOSE 9000
+
+CMD ["python", "/app/app.py"]
