--- requirements-all.txt
@@ -1,12 +1,0 @@
-numpy==1.26.4
-pandas==2.2.2
-paramiko==3.4.0
-Pillow==10.3.0
-plumbum==1.8.3
-PyYAML==6.0.1
-rich==13.7.1
-rpyc==6.0.0
-torch==2.3.0
-torchinfo==1.8.0
-torchvision==0.18.0
-ultralytics==8.2.25
