--- requirements-api.txt
@@ -1,6 +1,0 @@
-pandas==2.2.2
-paramiko==3.4.0
-plumbum==1.8.3
-PyYAML==6.0.1
-rich==13.7.1
-rpyc==6.0.0
