--- requirements-experiment.txt
@@ -1,10 +1,0 @@
-numpy==1.26.4
-pandas==2.2.2
-Pillow==10.3.0
-PyYAML==6.0.1
-rpyc==6.0.0
-# must be installed with cuda support
-# torch==2.3.0
-# torchinfo==1.8.0
-# torchvision==0.18.0
-# ultralytics==8.2.25
