+++ src/tracr/__init__.py
@@ -1,0 +1,6 @@
+from . import app_api
+from . import experiment_design
+
+__all__ = ["app_api", "experiment_design"]
+
+__version__ = "0.1.0"
