+++ src/tracr/app_api/device_mgmt.py
@@ -8,7 +8,7 @@ from concurrent.futures import ThreadPoolExecutor
 from queue import Queue
 from typing import Union
 
-from src.app_api import utils
+from src.tracr.app_api import utils
 
 
 class SSHAuthenticationException(Exception):
