+++ src/tracr/experiment_design/datasets/dataset.py
@@ -1,7 +1,7 @@
 from pathlib import Path
 from torch.utils.data import Dataset
 
-from src.app_api.utils import get_repo_root
+from src.tracr.app_api.utils import get_repo_root
 
 
 class BaseDataset(Dataset):
