+++ src/tracr/experiment_design/models/model_config.py
