+++ src/tracr/experiment_design/models/model_selector.py
