+++ src/tracr/experiment_design/partitioners/linreg_partitioner.py
